# docs/architecture/api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Daily Stock Analysis API",
    "description": "KR/US Watchlist AI Stock Analysis API\n\n## Modules\n- Stock analysis: trigger AI analysis\n- History: query historical analysis reports\n- Stock data: fetch market quotes\n\n## Authentication\nNo authentication required in the current version",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login or set initial password",
        "description": "Verify password and set session cookie. If password not set yet, accepts password+passwordConfirm.",
        "operationId": "auth_login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Change password",
        "description": "Change password. Requires valid session.",
        "operationId": "auth_change_password_api_v1_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/backtest/run": {
      "post": {
        "tags": [
          "Backtest"
        ],
        "summary": "백테스트 실행",
        "description": "히스토리 분석 기록을 백테스트 평가하고 backtest_results/backtest_summaries에 기록합니다.",
        "operationId": "run_backtest_api_v1_backtest_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "백테스트 실행 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestRunResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/results": {
      "get": {
        "tags": [
          "Backtest"
        ],
        "summary": "백테스트 결과 조회",
        "description": "백테스트 결과를 페이지 단위로 조회하며 종목 코드로 필터링할 수 있습니다.",
        "operationId": "get_backtest_results_api_v1_backtest_results_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "종목 코드 필터",
              "title": "Code"
            },
            "description": "종목 코드 필터"
          },
          {
            "name": "eval_window_days",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 120,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "평가 기간 필터",
              "title": "Eval Window Days"
            },
            "description": "평가 기간 필터"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "페이지 번호",
              "default": 1,
              "title": "Page"
            },
            "description": "페이지 번호"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "페이지당 항목 수",
              "default": 20,
              "title": "Limit"
            },
            "description": "페이지당 항목 수"
          }
        ],
        "responses": {
          "200": {
            "description": "백테스트 결과 목록",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResultsResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/performance": {
      "get": {
        "tags": [
          "Backtest"
        ],
        "summary": "전체 백테스트 성과 조회",
        "operationId": "get_overall_performance_api_v1_backtest_performance_get",
        "parameters": [
          {
            "name": "eval_window_days",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 120,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "평가 기간 필터",
              "title": "Eval Window Days"
            },
            "description": "평가 기간 필터"
          }
        ],
        "responses": {
          "200": {
            "description": "전체 백테스트 성과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceMetrics"
                }
              }
            }
          },
          "404": {
            "description": "백테스트 요약 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/performance/{code}": {
      "get": {
        "tags": [
          "Backtest"
        ],
        "summary": "개별 종목 백테스트 성과 조회",
        "operationId": "get_stock_performance_api_v1_backtest_performance__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "eval_window_days",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 120,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "평가 기간 필터",
              "title": "Eval Window Days"
            },
            "description": "평가 기간 필터"
          }
        ],
        "responses": {
          "200": {
            "description": "개별 종목 백테스트 성과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceMetrics"
                }
              }
            }
          },
          "404": {
            "description": "백테스트 요약 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BacktestResultsResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BacktestResultItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "limit"
        ],
        "title": "BacktestResultsResponse"
      },
      "BacktestRunRequest": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code",
            "description": "특정 종목만 백테스트합니다. 예: 005930 또는 AAPL"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "description": "기존 결과가 있어도 다시 계산합니다.",
            "default": false
          },
          "eval_window_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Eval Window Days",
            "description": "평가 기간(거래일 기준)"
          },
          "min_age_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 365.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Age Days",
            "description": "분석 기록의 최소 경과일(0=제한 없음)"
          },
          "limit": {
            "type": "integer",
            "maximum": 2000.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "처리할 분석 기록의 최대 개수",
            "default": 200
          }
        },
        "type": "object",
        "title": "BacktestRunRequest"
      },
      "BacktestRunResponse": {
        "properties": {
          "processed": {
            "type": "integer",
            "title": "Processed",
            "description": "후보 기록 수"
          },
          "saved": {
            "type": "integer",
            "title": "Saved",
            "description": "저장된 백테스트 결과 수"
          },
          "completed": {
            "type": "integer",
            "title": "Completed",
            "description": "완료된 백테스트 수"
          },
          "insufficient": {
            "type": "integer",
            "title": "Insufficient",
            "description": "데이터 부족 건수"
          },
          "errors": {
            "type": "integer",
            "title": "Errors",
            "description": "오류 건수"
          }
        },
        "type": "object",
        "required": [
          "processed",
          "saved",
          "completed",
          "insufficient",
          "errors"
        ],
        "title": "BacktestRunResponse"
      },
...
      "ChangePasswordRequest": {
        "properties": {
          "currentPassword": {
            "type": "string",
            "title": "Currentpassword",
            "default": ""
          },
          "newPassword": {
            "type": "string",
            "title": "Newpassword",
            "default": ""
          },
          "newPasswordConfirm": {
            "type": "string",
            "title": "Newpasswordconfirm",
            "default": ""
          }
        },
        "type": "object",
        "title": "ChangePasswordRequest",
        "description": "Change password request body."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error type",
            "example": "validation_error"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Error detail",
            "example": "요청 파라미터 오류"
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Additional error information"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Error response.",
        "example": {
          "error": "not_found",
          "message": "리소스를 찾을 수 없습니다"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Admin password",
            "default": ""
          },
          "passwordConfirm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passwordconfirm",
            "description": "Confirm (first-time)"
          }
        },
        "type": "object",
        "title": "LoginRequest",
        "description": "Login request body. For first-time setup use password + password_confirm."
      },
...
      "PerformanceMetrics": {
        "properties": {
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "eval_window_days": {
            "type": "integer",
            "title": "Eval Window Days"
          },
          "engine_version": {
            "type": "string",
            "title": "Engine Version"
          },
          "computed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Computed At"
          },
          "total_evaluations": {
            "type": "integer",
            "title": "Total Evaluations"
          },
          "completed_count": {
            "type": "integer",
            "title": "Completed Count"
          },
          "insufficient_count": {
            "type": "integer",
            "title": "Insufficient Count"
          },
          "long_count": {
            "type": "integer",
            "title": "Long Count"
          },
          "cash_count": {
            "type": "integer",
            "title": "Cash Count"
          },
          "win_count": {
            "type": "integer",
            "title": "Win Count"
          },
          "loss_count": {
            "type": "integer",
            "title": "Loss Count"
          },
          "neutral_count": {
            "type": "integer",
            "title": "Neutral Count"
          },
          "direction_accuracy_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction Accuracy Pct"
          },
          "win_rate_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Win Rate Pct"
          },
          "neutral_rate_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neutral Rate Pct"
          },
          "avg_stock_return_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Stock Return Pct"
          },
          "avg_simulated_return_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Simulated Return Pct"
          },
          "stop_loss_trigger_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss Trigger Rate"
          },
          "take_profit_trigger_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit Trigger Rate"
          },
          "ambiguous_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ambiguous Rate"
          },
          "avg_days_to_first_hit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Days To First Hit"
          },
          "advice_breakdown": {
            "additionalProperties": true,
            "type": "object",
            "title": "Advice Breakdown"
          },
          "diagnostics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Diagnostics"
          }
        },
        "type": "object",
        "required": [
          "scope",
          "eval_window_days",
          "engine_version",
          "total_evaluations",
          "completed_count",
          "insufficient_count",
          "long_count",
          "cash_count",
          "win_count",
          "loss_count",
          "neutral_count"
        ],
        "title": "PerformanceMetrics"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}